requirements.txt
Flask
mutagen
yt-dlp
pytube
spotipy
fuzzywuzzy
Pillow
requests
syncedlyrics
ytmusicapi
python-Levenshtein
python-dotenv
pycryptodome
cryptography
pytz